fix: recover working tree + session fixes after git-stash revert
Recovered from stash@{0} after a `git stash` + failed `git stash pop`
(blocked by root-owned Docker artifacts) reverted 25+ files to HEAD.

Session fixes, now durable:
- multi_index: overlay leg runs its OWN spot-adjustment (MCN SPOT_ADJ_FALL)
  + own-monthly-contract pick (drop fix); both legs adjust independently
- engine_rust: rel_leg_premium stepwise-from-ATM strike pick (no far-strike
  hunt); get_expiry_dates DataFrame misuse fixed (rel-leg PE no longer dropped)
- rust_fast_path: get_future_price ±1-day expiry tolerance (Bakri-Id
  holiday-split June-2023 futures no longer skip)
- excel_builder + backtest: midcap hypothetical Qty (capital-sized) shown in
  tradesheet, only when sizing is active

Excludes root-owned backend/native/target build artifacts.

+++ backend/services/algotest_job.py
@@ -437,7 +437,10 @@ def _try_rust_engine(payload, index, effective_from, effective_to):
                 # branches above (FUT via _fut_leg_mae_mfe and CE/PE via
                 # _calculate_leg_mae_mfe) return a plain unscaled ratio, so
                 # scale once here by this record's own lots (lot_size excluded).
-                _rec_lots = int(rec.get("lots") or 1)
+                # float(): integer lots are unchanged (1 → 1.0, ratio×1.0 == ratio×1)
+                # so existing runs are byte-identical, but a capital-sizing decimal
+                # qty scales MAE/MFE without truncation.
+                _rec_lots = float(rec.get("lots") or 1)
                 if mae_val is not None:
                     rec["MAE"] = mae_val * _rec_lots
                 if mfe_val is not None:
